+++ docs/CHANGELOG.md
@@ -25,6 +25,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   multiple affected[] blocks into a single range, preventing false-positive
   AFFECTED entries (OSIDB-5343)
 
+### Added
+- Certain workflow labels now provide a `reason` attribute which can provide
+  insights as to why it was applied (OSIDB-5318)
+
 ## [5.16.3] - 2026-08-13
 ### Fixed
 - fix disabled automatic timestamp update on task creation/update (OSIDB-5387)